clear GameScreen

+++ Rainbow/Modules/ResultsModule/Models/ResultModel.swift
@@ -6,9 +6,10 @@
 //
 
 import Foundation
+
 struct GameResultModel: Codable {
     var numberGame: Int
-    let durationGame: Double
+    let durationGame: Int
     let speedGame: Int
     let resultGame: String
 }
